refactor(proxy): move caching out of coco crate
Make the caching of announcements a session concern.

Closes #857
